Refactor new schema parsing code and add Schema subclasses
tagHasAttribute is now piped through Schema to help promote clarity
and avoid null-pointer errors.

As part of this commit, the dataset tests have been upgraded to the
stable v8.0.0 HED schema.
