+++ validator/schema/init.js
@@ -1,18 +1,19 @@
 const semver = require('semver')
 
-const schemaUtils = require('../../common/schema')
+const loadSchema = require('../../common/schema/loader')
+const { Schemas, Hed2Schema, Hed3Schema } = require('../../common/schema/types')
 
 const { buildMappingObject } = require('../../converter/schema')
 const { setParent } = require('../../utils/xml2js')
 
-const { Hed2SchemaParser, HedV8SchemaParser } = require('./parser')
-const { SchemaAttributes } = require('./types')
+const { Hed2SchemaParser } = require('./hed2')
+const { HedV8SchemaParser } = require('./hed3')
 
 /**
  * Build a schema attributes object from schema XML data.
  *
  * @param {object} xmlData The schema XML data.
- * @return {SchemaAttributes} The schema attributes object.
+ * @return {SchemaAttributes|SchemaEntries} The schema attributes object.
  */
 const buildSchemaAttributesObject = function (xmlData) {
   const rootElement = xmlData.HED
@@ -32,20 +33,21 @@ const buildSchemaAttributesObject = function (xmlData) {
  * @return {Promise<never>|Promise<Schemas>} The schema container object or an error.
  */
 const buildSchema = function (schemaDef = {}, useFallback = true) {
-  return schemaUtils.loadSchema(schemaDef, useFallback).then((xmlData) => {
+  return loadSchema(schemaDef, useFallback).then((xmlData) => {
     const schemaAttributes = buildSchemaAttributesObject(xmlData)
     const mapping = buildMappingObject(xmlData)
-    const baseSchema = new schemaUtils.Schema(
-      xmlData,
-      schemaAttributes,
-      mapping,
-    )
-    return new schemaUtils.Schemas(baseSchema)
+    const rootElement = xmlData.HED
+    let baseSchema
+    if (semver.gte(rootElement.$.version, '8.0.0-alpha.3')) {
+      baseSchema = new Hed3Schema(xmlData, schemaAttributes, mapping)
+    } else {
+      baseSchema = new Hed2Schema(xmlData, schemaAttributes, mapping)
+    }
+    return new Schemas(baseSchema)
   })
 }
 
 module.exports = {
   buildSchema: buildSchema,
   buildSchemaAttributesObject: buildSchemaAttributesObject,
-  SchemaAttributes: SchemaAttributes,
 }
